factor out wallet tokens hook
